requirements.txt
plotly==6.0.0
bs4==0.0.2
requests==2.31.0
requests-file==2.1.0
mojito2==0.1.6
wordcloud==1.9.4
streamlit==1.36.0
pandas==2.2.2
finance-datareader==0.9.96
supabase==2.13.0
python-dotenv==1.0.1